Move Controller logic to MainController
