+++ EvoCarcassonne/MainWindow.xaml
@@ -4,7 +4,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:EvoCarcassonne"
-        xmlns:views = "clr-namespace:EvoCarcassonne.View"
+        xmlns:views="clr-namespace:EvoCarcassonne.View"
         mc:Ignorable="d"
         Title="EvoCarcassonne" Height="450" Width="800">
     <Grid Margin="10">
